Refactor theme preference handling and chat toggle
